Split transactions to connectors and transformers.
