+++ transactron/lib/__init__.py
@@ -2,7 +2,6 @@ from .connectors import *  # noqa: F401
 from .buttons import *  # noqa: F401
 from .adapters import *  # noqa: F401
 from .transformers import *  # noqa: F401
-from .transactions import *  # noqa: F401
 from .reqres import *  # noqa: F401
 from .storage import *  # noqa: F401
 from .simultaneous import *  # noqa: F401
