+++ transactron/lib/connectors.py
@@ -6,6 +6,8 @@ __all__ = [
     "FIFO",
     "Forwarder",
     "Connect",
+    "ConnectTrans",
+    "ManyToOneConnectTrans",
 ]
 
 
@@ -182,5 +184,69 @@ class Connect(Elaboratable):
         def _(arg):
             m.d.av_comb += rev_read_value.eq(arg)
             return read_value
+
+        return m
+
+
+class ConnectTrans(Elaboratable):
+    """Simple connecting transaction.
+
+    Takes two methods and creates a transaction which calls both of them.
+    Result of the first method is connected to the argument of the second,
+    and vice versa. Allows easily connecting methods with compatible
+    layouts.
+    """
+
+    def __init__(self, method1: Method, method2: Method):
+        """
+        Parameters
+        ----------
+        method1: Method
+            First method.
+        method2: Method
+            Second method.
+        """
+        self.method1 = method1
+        self.method2 = method2
+
+    def elaborate(self, platform):
+        m = TModule()
+
+        with Transaction().body(m):
+            data1 = Record.like(self.method1.data_out)
+            data2 = Record.like(self.method2.data_out)
+
+            m.d.top_comb += data1.eq(self.method1(m, data2))
+            m.d.top_comb += data2.eq(self.method2(m, data1))
+
+        return m
+
+
+class ManyToOneConnectTrans(Elaboratable):
+    """Many-to-one method connection.
+
+    Connects each of a set of methods to another method using separate
+    transactions. Equivalent to a set of `ConnectTrans`.
+    """
+
+    def __init__(self, *, get_results: list[Method], put_result: Method):
+        """
+        Parameters
+        ----------
+        get_results: list[Method]
+            Methods to be connected to the `put_result` method.
+        put_result: Method
+            Common method for each of the connections created.
+        """
+        self.get_results = get_results
+        self.m_put_result = put_result
+
+        self.count = len(self.get_results)
+
+    def elaborate(self, platform):
+        m = TModule()
+
+        for i in range(self.count):
+            m.submodules[f"ManyToOneConnectTrans_input_{i}"] = ConnectTrans(self.m_put_result, self.get_results[i])
 
         return m
